+++ src/Instance/InstanceMovement/InstanceMovementDetails/InstanceMovementDetails.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { screen } from '@folio/jest-config-stripes/testing-library/react';
+import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 import { QueryClient, QueryClientProvider } from 'react-query';
 
 import '../../../../test/jest/__mock__';