+++ src/Instance/InstanceMovement/InstanceMovementDetails/InstanceMovementDetailsActions.test.js
@@ -1,8 +1,8 @@
 import '../../../../test/jest/__mock__';
 
 import React from 'react';
-import { screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { screen } from '@folio/jest-config-stripes/testing-library/react';
+import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
 import InstanceMovementDetailsActions from './InstanceMovementDetailsActions';
 import {