+++ src/ViewInstance.test.js
@@ -1,7 +1,7 @@
 import '../test/jest/__mock__';
 import React from 'react';
-import { screen, waitFor } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { screen, waitFor } from '@folio/jest-config-stripes/testing-library/react';
+import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { Router } from 'react-router-dom';
 import { createMemoryHistory } from 'history';
@@ -28,7 +28,7 @@ jest.mock('./components/ImportRecordModal/ImportRecordModal', () => (props) => {
       externalIdentifier: 'externalIdentifier',
       selectedJobProfileId: 'profileId'
     };
-    const container = 
+    const container =
     <div>
       <div>ImportRecordModal</div>
       <button type="button" onClick={() => handleSubmit(args)}>handleSubmit</button>