+++ src/components/NewOrderModal/NewOrderModalContainer.test.js
@@ -3,8 +3,8 @@ import {
   BrowserRouter as Router,
   useHistory,
 } from 'react-router-dom';
-import { act, render, screen } from '@testing-library/react';
-import user from '@testing-library/user-event';
+import { act, render, screen } from '@folio/jest-config-stripes/testing-library/react';
+import user from '@folio/jest-config-stripes/testing-library/user-event';
 
 import '../../../test/jest/__mock__';