+++ src/views/ItemView.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { noop } from 'lodash';
-import { waitFor, screen } from '@testing-library/react';
+import { waitFor, screen } from '@folio/jest-config-stripes/testing-library/react';
 
 
 import '../../test/jest/__mock__';